Issue #44: Added ColumnAggregationStep which executes fully on query remotes and aggregates various columns of a single row into a new value. Added syntax that allows queries to reference "all" elements of a repeated field (a[*].b for the b fields of all a elements) and used that in ColumnAggregationStep. Added test on Long columns. Adjusted AggregationFunction.ValueProvider in a way so it can be used for both, row aggregations (GROUP BY) and for the new col aggregations. Extended Request objects to support AGGREGATION_COL an the planner classes accordingly. All aggregation functions that have a parameter can now be used to aggregate on cols instead of rows, if the parameter contains the "all" syntax, example: avg(b[*].c).

+++ diqube-function/src/main/java/org/diqube/function/aggregate/AvgLongFunction.java
@@ -78,7 +78,7 @@ public class AvgLongFunction implements AggregationFunction<Long, IntermediaryRe
 
   @Override
   public ColumnType getOutputType() {
-    return ColumnType.LONG;
+    return ColumnType.DOUBLE;
   }
 
   @Override
